Crate script_extractor

Library to parse movie scripts and output them in a structured format

Modules

parse
Parsing scripts into Scripts.

serialize
Serialize Scripts into different formats

Structs

Location
Represents a location in which part of a Scene takes place.

Enums

DialogPart
The different parts of a Dialog.

LocationType
Represents the types of locations often used in scripts.

ScenePart
A Scene consists of Directions and Dialogs.

Functions

extract_range
Parses the given string into a range.

filter_script
Filter the script using a range of pages.

Type Definitions

Scene
A Scene consists of a list of Locations.

Script
A Script consists of a list of Scenes.