Module script_extractor::parse

Parsing scripts into Scripts.

This module provides functions to parse scripts which have been extracted from pdfs into Scripts.

Structs

Location
Represents a location in which part of a Scene takes place.

Enums

DialogPart
The different parts of a Dialog.

LocationType
Represents the types of locations often used in scripts.

ScenePart
A Scene consists of Directions and Dialogs.

Functions

parse_script
Parses the given script into a Script.

Type Definitions

Scene
A Scene consists of a list of Locations.

Script
A Script consists of a list of Scenes.