Function script_extractor::serialize::json::format_script

pub fn format_script<W: Write>(scenes: &Script, output: &mut W) -> EncodeResult<()>

Serialize the given Script into a json

Serialize the given Script into the following json format and write it to the given Writer.

Example

[
    [
        {
            "place": "Snowy Landscape",
            "type": "external",
            "parts": [{
                "page": 1,
                "direction": "Swirls of snow obscure the rocky formations of a mountain (...)"
            },{
                "page": 1,
                "direction": "Five ragged men attack a young girl, SINTEL, (...)"
            },{
                "page": 1,
                "character": "Shaman",
                "dialog": [
                    "You’re lucky to be alive. (...)"
                ]
            },{
                "page": 2,
                "direction": "Finally she collapses into the snow, her eyes shut tight."
            },{
                "page": 2,
                "direction": "BLACK"
            },{
                "page": 2,
                "character": "Shaman",
                "mode": "VO",
                "dialog": [
                    { "direction": "(To Sintel)" },
                    "Here, take a sip."
                ]
            }]
        }
    ]
]