Function script_extractor::serialize::xml::format_script

pub fn format_script<W: Write>(scenes: &Script, output: &mut W) -> XmlResult<()>

Serialize the given Script into a xml

Serialize the given Script into the following xml format and write it to the given Writer.

Example

<?xml version="1.0" encoding="UTF-8"?>
<script>
  <scene>
    <location place="Snowy Landscape" type="external">
      <direction page="1">Swirls of snow obscure the rocky formations of a mountain (...)</direction>
      <direction page="1">Five ragged men attack a young girl, SINTEL, (...)</direction>
      <!-- (...) -->
      <dialog character="Shaman" page="1">You’re lucky to be alive. (...)</dialog>
      <direction page="2">Finally she collapses into the snow, her eyes shut tight.</direction>
      <direction page="2">BLACK</direction>
      <dialog character="Shaman" mode="VO" page="2">Here, take a sip.</dialog>
    </location>
  </scene>
  <!-- (...) -->
</script>